+++ apps/picsa-tools/resources-tool/src/app/services/resources-tool.service.ts
@@ -278,77 +278,4 @@ export class ResourcesToolService extends PicsaAsyncService {
       // await Filesystem.deleteFile({ path: cacheFileUri, directory: Directory.Cache });
     }
   }
-  private thumbnailCache: Map<string, string> = new Map();
-
-  public async generateVideoThumbnail(videoUrl: string, captureTime = 20): Promise<string> {
-    if (this.thumbnailCache.has(videoUrl)) {
-      return Promise.resolve(this.thumbnailCache.get(videoUrl)!);
-    }
-    try {
-      return await new Promise<string>((resolve, reject) => {
-        const video = document.createElement('video');
-        video.crossOrigin = 'anonymous';
-        video.muted = true;
-        video.playsInline = true;
-        video.src = videoUrl;
-
-        video.addEventListener(
-          'loadedmetadata',
-          () => {
-            if (video.duration < captureTime) {
-              captureTime = video.duration / 2;
-            }
-            video.currentTime = captureTime;
-          },
-          { once: true }
-        );
-
-        video.addEventListener(
-          'seeked',
-          () => {
-            const canvas = document.createElement('canvas');
-            canvas.width = video.videoWidth;
-            canvas.height = video.videoHeight;
-
-            const ctx = canvas.getContext('2d');
-            if (!ctx) {
-              console.error('Unable to get canvas context.');
-              return resolve('');
-            }
-
-            ctx.drawImage(video, 0, 0, canvas.width, canvas.height);
-            const dataUrl = canvas.toDataURL('image/webp');
-
-            // free up memory
-            if (videoUrl.startsWith('blob:')) {
-              URL.revokeObjectURL(videoUrl);
-            }
-            this.thumbnailCache.set(videoUrl, dataUrl);
-
-            resolve(dataUrl);
-          },
-          { once: true }
-        );
-
-        video.addEventListener(
-          'error',
-          (err) => {
-            console.error('Error loading video:', err);
-
-            if (videoUrl.startsWith('blob:')) {
-              URL.revokeObjectURL(videoUrl);
-            }
-            // Return '' instead of rejecting
-            resolve('');
-          },
-          { once: true }
-        );
-
-        video.load();
-      });
-    } catch (err) {
-      console.error('Unexpected error generating thumbnail:', err);
-      return '';
-    }
-  }
 }
