+++ libs/shared/src/features/video-player/index.ts
@@ -5,11 +5,12 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatProgressBarModule } from '@angular/material/progress-bar';
 
 import { VideoPlayerComponent } from './video-player.component';
+import { VideoThumbnailComponent } from './video-thumbnail/video-thumbnail.component';
 
 @NgModule({
   imports: [CommonModule, MatButtonModule, MatIconModule, MatProgressBarModule],
   exports: [VideoPlayerComponent],
-  declarations: [VideoPlayerComponent],
+  declarations: [VideoPlayerComponent, VideoThumbnailComponent],
   providers: [],
 })
 /**
