+++ libs/shared/src/features/video-player/video-thumbnail/video-thumbnail.component.html
@@ -1,0 +1,12 @@
+<div>
+    <div 
+        class="thumbnail" 
+        [style.background-image]="'url(' + (thumbnail || generatedThumbnail) + ')'"
+        *ngIf="thumbnail || generatedThumbnail">
+    </div>
+    <div 
+        class="thumbnail-status" 
+        *ngIf="!thumbnail && generatedThumbnail === undefined">
+        loading...
+    </div>
+</div>
